add AbstractBackendHandler
